+++ pos_analytic_tag/doc/RELEASE_NOTES.md
@@ -4,3 +4,11 @@
 #### Version 17.0.1.0.0
 #### ADD
  - Initial Commit for PoS Analytic Tag
+
+## Module <pos_analytic_tag>
+
+#### 19.06.2024
+#### Version 17.0.1.0.1
+#### BUGFIX
+
+- Updated the bug with analytic tag of the journal entries